refactor(arco-bench): split benchmark runner into modules
